requirements.txt
whisper
ffmpeg-python
numpy
opencv-python-headless
openai-whisper
moviepy
numpy
torch